Refactored URLVoid to support APIVoid (vendor migration), Updated readme with new APIVoid URL and name, Updated apikey.yaml with new name and to remove unused identifier, Updated readme with change history, Updated Phishtank URL to HTTPS, Removed unused imports
